Text.Megaparsec.ByteString

Description

Convenience definitions for working with strict ByteString.

Documentation

type Parser = Parsec Dec ByteString

Modules corresponding to various types of streams define Parser accordingly, so the user can use it to easily change type of input stream by importing different “type modules”. This one is for strict ByteStrings.