Text.Megaparsec.Error

Description

Parse errors. The current version of Megaparsec supports typed errors instead of String-based ones. This gives a lot of flexibility in describing what exactly went wrong as well as a way to return arbitrary data in case of failure.

You probably do not want to import this module directly because Text.Megaparsec re-exports it anyway.

Parse error type

data ErrorItem t

A data type that is used to represent “unexpected/expected” items in ParseError. It is parametrized over the token type t.

Since: 5.0.0

Constructors

Tokens (NonEmpty t)
Non-empty stream of tokens

Label (NonEmpty Char)
Label (cannot be empty)

EndOfInput
End of input

data ErrorFancy e

Additional error data, extendable by user. When no custom data is necessary, the type is typically indexed by Void to “cancel” the ErrorCustom constructor.

Since: 6.0.0

Constructors

ErrorFail String
fail has been used in parser monad

ErrorIndentation Ordering Pos Pos
Incorrect indentation error: desired ordering between reference level and actual level, reference indentation level, actual indentation level

ErrorCustom e
Custom error data

data ParseError s e

ParseError s e represents a parse error parametrized over the stream type s and the custom data e.

Semigroup and Monoid instances of the data type allow us to merge parse errors from different branches of parsing. When merging two ParseErrors, the longest match is preferred; if positions are the same, custom data sets and collections of message items are combined. Note that fancy errors take precedence over trivial errors in merging.

Since: 7.0.0

Constructors

TrivialError Int (Maybe (ErrorItem (Token s))) (Set (ErrorItem (Token s)))
Trivial errors, generated by the Megaparsec's machinery. The data constructor includes the offset of error, unexpected token (if any), and expected tokens.

Type of the first argument was changed in the version 7.0.0.

FancyError Int (Set (ErrorFancy e))
Fancy, custom errors.

Type of the first argument was changed in the version 7.0.0.

mapParseError :: Ord e' => (e -> e') -> ParseError s e -> ParseError s e'

Modify the custom data component in a parse error. This could be done via fmap if not for the Ord constraint.

Since: 7.0.0

errorOffset :: ParseError s e -> Int

Get the offset of a ParseError.

Since: 7.0.0

setErrorOffset :: Int -> ParseError s e -> ParseError s e

Set the offset of a ParseError.

Since: 8.0.0

data ParseErrorBundle s e

A non-empty collection of ParseErrors equipped with PosState that allows us to pretty-print the errors efficiently and correctly.

Since: 7.0.0

Constructors

ParseErrorBundle

Fields

bundleErrors :: NonEmpty (ParseError s e)

A collection of ParseErrors that is sorted by parse error offsets

bundlePosState :: PosState s

The state that is used for line/column calculation

attachSourcePos

Arguments

:: (Traversable t, TraversableStream s)

=> (a -> Int)
How to project offset from an item (e.g. errorOffset)

-> t a
The collection of items

-> PosState s
Initial PosState

-> (t (a, SourcePos), PosState s)
The collection with SourcePoses added and the final PosState

Attach SourcePoses to items in a Traversable container given that there is a projection allowing us to get an offset per item.

Items must be in ascending order with respect to their offsets.

Since: 7.0.0

Pretty-printing

class Ord a => ShowErrorComponent a where

The type class defines how to print a custom component of ParseError.

Since: 5.0.0

Minimal complete definition

showErrorComponent

Methods

showErrorComponent :: a -> String

Pretty-print a component of ParseError.

errorComponentLen :: a -> Int

Length of the error component in characters, used for highlighting of parse errors in input string.

Since: 7.0.0

errorBundlePretty

Arguments

:: forall s e. (VisualStream s, TraversableStream s, ShowErrorComponent e)

=> ParseErrorBundle s e
Parse error bundle to display

-> String
Textual rendition of the bundle

Pretty-print a ParseErrorBundle. All ParseErrors in the bundle will be pretty-printed in order together with the corresponding offending lines by doing a single pass over the input stream. The rendered String always ends with a newline.

Since: 7.0.0

parseErrorPretty

Arguments

:: (VisualStream s, ShowErrorComponent e)

=> ParseError s e
Parse error to render

-> String
Result of rendering

Pretty-print a ParseError. The rendered String always ends with a newline.

Since: 5.0.0

parseErrorTextPretty

Arguments

:: forall s e. (VisualStream s, ShowErrorComponent e)

=> ParseError s e
Parse error to render

-> String
Result of rendering

Pretty-print a textual part of a ParseError, that is, everything except for its position. The rendered String always ends with a newline.

Since: 5.1.0

showErrorItem :: VisualStream s => Proxy s -> ErrorItem (Token s) -> String

Pretty-print an ErrorItem.

Since: 9.4.0