Text.Megaparsec.Error

Description

Parse errors. Current version of Megaparsec supports well-typed errors instead of String-based ones. This gives a lot of flexibility in describing what exactly went wrong as well as a way to return arbitrary data in case of failure.

Documentation

data ErrorItem t

Data type that is used to represent “unexpected/expected” items in ParseError. The data type is parametrized over the token type t.

Since: 5.0.0

Constructors

Tokens (NonEmpty t)
Non-empty stream of tokens

Label (NonEmpty Char)
Label (cannot be empty)

EndOfInput
End of input

class Ord e => ErrorComponent e where

The type class defines how to represent information about various exceptional situations. Data types that are used as custom data component in ParseError must be instances of this type class.

Since: 5.0.0

Minimal complete definition

representFail, representIndentation

Methods

representFail :: String -> e

Represent the message passed to fail in parser monad.

Since: 5.0.0

representIndentation :: Ordering -> Pos -> Pos -> e

Represent information about incorrect indentation.

Since: 5.0.0

data Dec

“Default error component”. This is our instance of ErrorComponent provided out-of-box.

Since: 5.0.0

Constructors

DecFail String
fail has been used in parser monad

DecIndentation Ordering Pos Pos
Incorrect indentation error: desired ordering between reference level and actual level, reference indentation level, actual indentation level

data ParseError t e

ParseError represents… parse errors. It provides the stack of source positions, a set of expected and unexpected tokens as well as a set of custom associated data. The data type is parametrized over the token type t and the custom data e.

Note that the stack of source positions contains current position as its head, and the rest of positions allows to track full sequence of include files with topmost source file at the end of the list.

Semigroup (and Monoid) instance of the data type allows to merge parse errors from different branches of parsing. When merging two ParseErrors, the longest match is preferred; if positions are the same, custom data sets and collections of message items are combined.

Constructors

ParseError

Fields

errorPos :: NonEmpty SourcePos

Stack of source positions

errorUnexpected :: Set (ErrorItem t)

Unexpected items

errorExpected :: Set (ErrorItem t)

Expected items

errorCustom :: Set e

Associated data, if any

class ShowToken a where

Type class ShowToken includes methods that allow to pretty-print single token as well as stream of tokens. This is used for rendering of error messages.

Minimal complete definition

showTokens

Methods

showTokens :: NonEmpty a -> String

Pretty-print non-empty stream of tokens. This function is also used to print single tokens (represented as singleton lists).

Since: 5.0.0

class Ord a => ShowErrorComponent a where

The type class defines how to print custom data component of ParseError.

Since: 5.0.0

Minimal complete definition

showErrorComponent

Methods

showErrorComponent :: a -> String

Pretty-print custom data component of ParseError.

parseErrorPretty

Arguments

:: (Ord t, ShowToken t, ShowErrorComponent e)

=> ParseError t e
Parse error to render

-> String
Result of rendering

Pretty-print a ParseError. The rendered String always ends with a newline.

The function is defined as:

parseErrorPretty e =
  sourcePosStackPretty (errorPos e) ++ ":\n" ++ parseErrorTextPretty e

Since: 5.0.0

sourcePosStackPretty :: NonEmpty SourcePos -> String

Pretty-print a stack of source positions.

Since: 5.0.0

parseErrorTextPretty

Arguments

:: (Ord t, ShowToken t, ShowErrorComponent e)

=> ParseError t e
Parse error to render

-> String
Result of rendering

Pretty-print a textual part of a ParseError, that is, everything except stack of source positions. The rendered staring always ends with a new line.

Since: 5.1.0