Text.Megaparsec.Error.Builder

Description

A set of helpers that should make construction of ParseErrors more concise. This is primarily useful in test suites and for debugging.

Since: 6.0.0

Top-level helpers

err

Arguments

:: Int
ParseError offset

-> ET s
Error components

-> ParseError s e
Resulting ParseError

Assemble a ParseError from offset and ET t value. ET t is a monoid and can be assembled by combining primitives provided by this module, see below.

errFancy

Arguments

:: Int
ParseError offset

-> EF e
Error components

-> ParseError s e
Resulting ParseError

Like err, but constructs a “fancy” ParseError.

Error components

utok :: Stream s => Token s -> ET s

Construct an “unexpected token” error component.

utoks :: forall s. Stream s => Tokens s -> ET s

Construct an “unexpected tokens” error component. Empty chunk produces EndOfInput.

ulabel :: Stream s => String -> ET s

Construct an “unexpected label” error component. Do not use with empty strings (for empty strings it's bottom).

ueof :: Stream s => ET s

Construct an “unexpected end of input” error component.

etok :: Stream s => Token s -> ET s

Construct an “expected token” error component.

etoks :: forall s. Stream s => Tokens s -> ET s

Construct an “expected tokens” error component. Empty chunk produces EndOfInput.

elabel :: Stream s => String -> ET s

Construct an “expected label” error component. Do not use with empty strings.

eeof :: Stream s => ET s

Construct an “expected end of input” error component.

fancy :: ErrorFancy e -> EF e

Construct a custom error component.

Data types

data ET s

Auxiliary type for construction of trivial parse errors.

data EF e

Auxiliary type for construction of fancy parse errors.