Text.Megaparsec.Error.Builder

Description

A set of helpers that should make construction of ParseErrors more concise. This is primarily useful in test suites and for debugging, you most certainly don't need it for normal usage.

Since: 6.0.0

Top-level helpers

err

Arguments

:: NonEmpty SourcePos
ParseError position

-> ET t
Error components

-> ParseError t e
Resulting ParseError

Assemble a ParseError from source position and ET t value. To create source position, two helpers are available: posI and posN. ET t is a monoid and can be assembled by combining primitives provided by this module, see below.

errFancy

Arguments

:: NonEmpty SourcePos
ParseError position

-> EF e
Error components

-> ParseError t e
Resulting ParseError

Like err, but constructs a “fancy” ParseError.

Error position

posI :: NonEmpty SourcePos

Initial source position with empty file name.

posN :: forall s. Stream s => Int -> s -> NonEmpty SourcePos

posN n s returns source position achieved by applying advanceN method corresponding to the type of stream s.

Error components

utok :: Ord t => t -> ET t

Construct an “unexpected token” error component.

utoks :: Ord t => [t] -> ET t

Construct an “unexpected tokens” error component. Empty string produces EndOfInput.

ulabel :: Ord t => String -> ET t

Construct an “unexpected label” error component. Do not use with empty strings (for empty strings it's bottom).

ueof :: Ord t => ET t

Construct an “unexpected end of input” error component.

etok :: Ord t => t -> ET t

Construct an “expected token” error component.

etoks :: Ord t => [t] -> ET t

Construct an “expected tokens” error component. Empty string produces EndOfInput.

elabel :: Ord t => String -> ET t

Construct an “expected label” error component. Do not use with empty strings.

eeof :: Ord t => ET t

Construct an “expected end of input” error component.

fancy :: ErrorFancy e -> EF e

Construct a custom error component.

Data types

data ET t

Auxiliary type for construction of trivial parse errors.

data EF e

Auxiliary type for construction of fancy parse errors.