Text.Megaparsec.Internal

Description

Internal definitions. Versioning rules do not apply here. Please do not rely on these unless you really know what you're doing.

Since: 6.5.0

Data types

newtype Hints t

Hints represent a collection of ErrorItems to be included into ParseError (when it's a TrivialError) as “expected” message items when a parser fails without consuming input right after successful parser that produced the hints.

For example, without hints you could get:

>>> parseTest (many (char 'r') <* eof) "ra"
1:2:
unexpected 'a'
expecting end of input

We're getting better error messages with the help of hints:

>>> parseTest (many (char 'r') <* eof) "ra"
1:2:
unexpected 'a'
expecting 'r' or end of input

Constructors

Hints [Set (ErrorItem t)]

data Reply e s a

All information available after parsing. This includes consumption of input, success (with returned value) or failure (with parse error), and parser state at the end of parsing.

See also: Consumption, Result.

Constructors

Reply (State s) Consumption (Result s e a)

data Consumption

Whether the input has been consumed or not.

See also: Result, Reply.

Constructors

Consumed
Some part of input stream was consumed

Virgin
No input was consumed

data Result s e a

Whether the parser has failed or not. On success we include the resulting value, on failure we include a ParseError.

See also: Consumption, Reply.

Constructors

OK a
Parser succeeded

Error (ParseError s e)
Parser failed

newtype ParsecT e s m a

ParsecT e s m a is a parser with custom data component of error e, stream type s, underlying monad m and return type a.

Constructors

ParsecT

Fields

unParser :: forall b. State s -> (a -> State s -> Hints (Token s) -> m b) -> (ParseError s e -> State s -> m b) -> (a -> State s -> Hints (Token s) -> m b) -> (ParseError s e -> State s -> m b) -> m b

Helper functions

toHints

Arguments

:: Stream s

=> Int
Current offset in input stream

-> ParseError s e
Parse error to convert

-> Hints (Token s)

Convert ParseError record to Hints.

withHints

Arguments

:: Stream s

=> Hints (Token s)
Hints to use

-> (ParseError s e -> State s -> m b)
Continuation to influence

-> ParseError s e
First argument of resulting continuation

-> State s
Second argument of resulting continuation

-> m b

withHints hs c makes “error” continuation c use given hints hs.

Note that if resulting continuation gets ParseError that has custom data in it, hints are ignored.

accHints

Arguments

:: Hints t
Hints to add

-> (a -> State s -> Hints t -> m b)
An “OK” continuation to alter

-> a -> State s -> Hints t -> m b
Altered “OK” continuation

accHints hs c results in “OK” continuation that will add given hints hs to third argument of original continuation c.

refreshLastHint :: Hints t -> Maybe (ErrorItem t) -> Hints t

Replace the most recent group of hints (if any) with the given ErrorItem (or delete it if Nothing is given). This is used in the label primitive.

runParsecT

Arguments

:: Monad m

=> ParsecT e s m a
Parser to run

-> State s
Initial state

-> m (Reply e s a)

Low-level unpacking of the ParsecT type.

withParsecT :: (Monad m, Ord e') => (e -> e') -> ParsecT e s m a -> ParsecT e' s m a

Transform any custom errors thrown by the parser using the given function. Similar in function and purpose to withExceptT.

Since: 7.0.0