Text.Megaparsec.Pos

Description

Textual source position. The position includes name of file, line number, and column number. List of such positions can be used to model a stack of include files.

You probably do not want to import this module directly because Text.Megaparsec re-exports it anyway.

Abstract position

data Pos

Pos is the type for positive integers. This is used to represent line number, column number, and similar things like indentation level. Semigroup instance can be used to safely and efficiently add Poses together.

Since: 5.0.0

mkPos :: Int -> Pos

Construction of Pos from Int. The function throws InvalidPosException when given a non-positive argument.

Since: 6.0.0

unPos :: Pos -> Int

Extract Int from Pos.

Since: 6.0.0

pos1 :: Pos

Position with value 1.

Since: 6.0.0

defaultTabWidth :: Pos

Value of tab width used by default. Always prefer this constant when you want to refer to the default tab width because actual value may change in future.

Since: 5.0.0

data InvalidPosException

The exception is thrown by mkPos when its argument is not a positive number.

Since: 5.0.0

Constructors

InvalidPosException Int
The first value is the minimal allowed value, the second value is the actual value that was passed to mkPos.

Source position

data SourcePos

The data type SourcePos represents source positions. It contains the name of the source file, a line number, and a column number. Source line and column positions change intensively during parsing, so we need to make them strict to avoid memory leaks.

Constructors

SourcePos

Fields

sourceName :: FilePath

Name of source file

sourceLine :: !Pos

Line number

sourceColumn :: !Pos

Column number

initialPos :: FilePath -> SourcePos

Construct initial position (line 1, column 1) given name of source file.

sourcePosPretty :: SourcePos -> String

Pretty-print a SourcePos.

Since: 5.0.0