Text.Megaparsec.State

Description

Definition of Megaparsec's State.

Since: 6.5.0

Documentation

data State s e

This is the Megaparsec's state parametrized over stream type s and custom error component type e.

Constructors

State

Fields

stateInput :: s

The rest of input to process

stateOffset :: !Int

Number of processed tokens so far

Since: 7.0.0

statePosState :: PosState s

State that is used for line/column calculation

Since: 7.0.0

stateParseErrors :: [ParseError s e]

Collection of “delayed” ParseErrors in reverse order. This means that the last registered error is the first element of the list.

Since: 8.0.0

initialState

Arguments

:: FilePath
Name of the file the input is coming from

-> s
Input

-> State s e

Given the name of the source file and the input construct the initial state for a parser.

Since: 9.6.0

data PosState s

A special kind of state that is used to calculate line/column positions on demand.

Since: 7.0.0

Constructors

PosState

Fields

pstateInput :: s

The rest of input to process

pstateOffset :: !Int

Offset corresponding to beginning of pstateInput

pstateSourcePos :: !SourcePos

Source position corresponding to beginning of pstateInput

pstateTabWidth :: Pos

Tab width to use for column calculation

pstateLinePrefix :: String

Prefix to prepend to offending line

initialPosState

Arguments

:: FilePath
Name of the file the input is coming from

-> s
Input

-> PosState s

Given the name of source file and the input construct the initial positional state.

Since: 9.6.0